clean - cache auth
