+++ server/src/Cache/authentication/getAuthCache.ts
@@ -1,0 +1,8 @@
+import cache from '../'
+import getCacheNameByUserID from './getCacheNameByUserID';
+
+const getAuthCache = (userId : string) => {
+    return cache.get(getCacheNameByUserID(userId));
+}
+
+export default getAuthCache;
