+++ server/src/Cache/authentication/getCacheNameByUserID.ts
@@ -1,0 +1,7 @@
+const AUTHENTICATION_PREFIX = '/auth/';
+
+const getCacheNameByUserID = ( userId : string ) => {
+    return AUTHENTICATION_PREFIX + userId
+}
+
+export default getCacheNameByUserID;
