+++ server/src/Cache/authentication/hasAuthCache.ts
@@ -1,0 +1,8 @@
+import cache from '../';
+import getCacheNameByUserID from './getCacheNameByUserID'
+
+const hasAuthCache = ( userId : string ) => {
+    return cache.has(getCacheNameByUserID(userId))
+}
+
+export default hasAuthCache;
