+++ server/src/Cache/authentication/removeAuthCache.ts
@@ -1,0 +1,8 @@
+import cache from '../'
+import getCacheNameByUserID from './getCacheNameByUserID';
+
+const removeAuthCache = (userId : string) => {
+    cache.del(getCacheNameByUserID(userId));
+}
+
+export default removeAuthCache;
