+++ server/src/Cache/authentication/setAuthCache.ts
@@ -1,0 +1,10 @@
+import cache from '../'
+import getCacheNameByUserID from './getCacheNameByUserID'
+
+const AUTH_TTL = 60;
+
+const setAuthCache = ( userId : string, data : any ) => {
+    cache.set(getCacheNameByUserID(userId), data , AUTH_TTL)
+}
+
+export default setAuthCache;
